+++ src/test/resources/apis/v2/retry.json
